refactor(tagbookmark): BookmarkController 메서드 이동 및 테스트수정
