+++ backend/core/src/main/java/com/woowacourse/taggle/tag/service/BookmarkService.java
@@ -9,7 +9,6 @@ import com.woowacourse.taggle.tag.domain.Bookmark;
 import com.woowacourse.taggle.tag.domain.BookmarkRepository;
 import com.woowacourse.taggle.tag.dto.BookmarkCreateDto;
 import com.woowacourse.taggle.tag.dto.BookmarkResponse;
-import com.woowacourse.taggle.tag.dto.BookmarkTagResponse;
 import com.woowacourse.taggle.tag.exception.BookmarkNotFoundException;
 import com.woowacourse.taggle.user.domain.User;
 import com.woowacourse.taggle.user.dto.SessionUser;
@@ -32,13 +31,6 @@ public class BookmarkService {
                 .orElse(bookmarkRepository.save(bookmarkCreateDto.toEntityWithUser(user)));
 
         return BookmarkResponse.of(bookmark);
-    }
-
-    @Transactional(readOnly = true)
-    public BookmarkTagResponse findBookmark(final SessionUser user, final Long id) {
-        final Bookmark bookmark = findByIdAndUserId(id, user.getId());
-
-        return BookmarkTagResponse.of(bookmark);
     }
 
     @Transactional(readOnly = true)
