+++ backend/web/src/test/java/com/woowacourse/taggle/tag/controller/BookmarkControllerTest.java
@@ -16,7 +16,6 @@ import com.woowacourse.taggle.setup.domain.TagSetup;
 import com.woowacourse.taggle.setup.domain.UserSetup;
 import com.woowacourse.taggle.tag.controller.docs.BookmarkDocumentation;
 import com.woowacourse.taggle.tag.domain.Bookmark;
-import com.woowacourse.taggle.tag.domain.Tag;
 import com.woowacourse.taggle.user.domain.User;
 
 class BookmarkControllerTest extends ControllerTest {
@@ -46,19 +45,6 @@ class BookmarkControllerTest extends ControllerTest {
     void createBookmark() throws Exception {
         createByJsonParams(user, "/api/v1/bookmarks", "{ \"url\": \"http://github.com\" }")
                 .andDo(BookmarkDocumentation.createBookmark());
-    }
-
-    @WithMockUser(value = "ADMIN")
-    @DisplayName("findBookmark: 하나의 북마크를 조회한다.")
-    @Test
-    void findBookmark() throws Exception {
-        final Bookmark bookmark = bookmarkSetup.save(user);
-        final Tag tag = tagSetup.save(user);
-        tagBookmarkSetup.save(tag, bookmark);
-
-        readByPathVariables(user, "/api/v1/bookmarks/{bookmarkId}/tags", bookmark.getId())
-                .andExpect(jsonPath("$.id", is(bookmark.getId().intValue())))
-                .andDo(BookmarkDocumentation.findBookmark());
     }
 
     @WithMockUser(value = "ADMIN")
