+++ backend/web/src/test/java/com/woowacourse/taggle/tag/controller/docs/BookmarkDocumentation.java
@@ -21,21 +21,6 @@ public class BookmarkDocumentation {
         );
     }
 
-    public static RestDocumentationResultHandler findBookmark() {
-        return document("bookmarks/get",
-                pathParameters(
-                        parameterWithName("bookmarkId").description("북마크 ID")
-                ),
-                responseFields(
-                        fieldWithPath("id").description("북마크 ID"),
-                        fieldWithPath("url").description("북마크 URL"),
-                        fieldWithPath("tags").description("북마크의 태그 (Nullable)").optional(),
-                        fieldWithPath("tags.[].id").description("북마크에 있는 태그 ID"),
-                        fieldWithPath("tags.[].name").description("북마크에 있는 태그 이름")
-                )
-        );
-    }
-
     public static RestDocumentationResultHandler findBookmarks() {
         return document("bookmarks/get/List",
                 responseFields(
